Refactor the creation of LiveOS related images into a separate file.
